refactored App.jsx into modular components
